Firmware: Got a simple build running
Added a simple makefile which is capable of building C and C++ files.
Also added an empty main.cpp file to test those builds. All build
artifacts are stored in `firmware/bin` directory, which is also added
into .gitignore. Instructions for building and flashing are located in
the Readme of the firmware