Firmware: Added debug uart capabilities
Implemented one-way UART capability to the firmware that can be enabled
using a conditional `debuguart` flag passed to the makefile. The uart
can be used as a standalone output stream by calling the `uart::putc`
function or wired up to the standard output, thus giving the ability to
use standard functions such as `printf`. The intended use for this uart
currently is debug logging, thus additional macros were implemented
allowing conditional printing of debug logs.

The UART is hardcoded to be 8 data bits and 1 stop bit without parity.
Speed is configurable at compile time.