Firmware: Add display controller
Implemented display logic, which is responsible for interfacing with
the LC75823E radio faceplate using CCD protocol. The code in the
display controller takes over the SPI hardware of the microcontroller
and combines that with few pins to handle the physical layer of the
interface.

Data transfers are handled asynchronously via interrupts. The data that
is being transfered is double-buffered at the start of data transfer
(drawing) to enable modify-during-write functionality. This design was
chosen, because I have some spare bytes in the data memory.

An abstraction is also implemented which allows modifying individual
bits of the data array that gets sent to the display controller. Keep in
mind that addressing bytes outside of data array bounds is undefined
behavor.