Firmware: Add display segment addresses for icons
Implemented a header file `display_segments.hpp`, which has addresses of
each icons present on the display matrix of the radio faceplate. The
addresses are intended to be used with `set/clear/toggle_segment`
functions found in the `display.hpp` file and builds opon the
abstraction that is established in the file.