Firmware: Character map abstraction and symbols
Implemented a set of abstractions titled `segment_map_*` intended to
map bit masks into segments on a display. These abstractions come in two
flavors - single and multi.
Single segment maps are used to draw a pattern on a display where there
are segments only for that particular pattern (e.g. progress bars, the
single number indicating a radio station, CD spinner animation).
Multi segment maps are used to draw the main 8 13 segment characters
on the display. The map itself only maps segments for the first
character and then uses an algorithm to shift those numbers in the
correct position.

Due to memory constraints, bounds checking for segment map and shifting
algorithms were not implemented. It is up to the developer to ensure
that the boundaries are not exceeded. Doing out of boundary operations
is undefined behaviour.

As a demonstration for these maps, a single 7-segment digit (single map)
and 8 13 segment displays (multi map) were implemented. Bit maps for
characters that are intended to be displayed using their respective maps
were also defined.