Firmware: Add backlight control
Added the ability in the display controller to change the backlight
brightness. It was implemented using using Timer/Counter0 running in
fast PWM mode. The implementation only uses OCR0A register which maps to
PORTD6. PWM frequency is ~4kHz.

Two helper functions were exposed in the display controller header file
to allow users of the controller to set or get the value in range
[0:255].

It is undefined behavior to get/set brightness value if display
controller is not initialized.