Input: Rotary encoder support
Implemented support for the on-board rotary encoder. The rotary encoder
is connected via ADC (and a pull-up resistor) and has a distinctive
pattern (described in code) when CW or CCW rotation is performed. The
logic in input.cpp is used to detect this pattern and increment or
decrement an internal counter, thus keeping track of the encoder's
position.

This value is accessable via pop_rotary function found in input.hpp
file.