Buzzer: Add a module that plays some notes
Created a module named buzzer, that handles the external magnetic
resonator to play some notes. This module takes consumes two timers:
Timer1 and Timer2. The first one is used to generate the tone and the
second one is used to generate beeps for a certain amout of time. The
public API also exposes functions to continously play a note.

Values for the tone generation timer are precomputed based on an 8 note
scale, ranging from C to A, with C being 2048Hz - the resonant frequency
of the buzzer.