Input: Maskable and raw buttons
Added the ability to mask certain buttons in input controller. When
masked, the button is exempt from processing and beeping on press.
Up/Down buttons are important for screen switching, thus they are exempt
from masking.

Also implemented raw button helpers, which let a caller retrieve
unprocessed button statues. This is used when more fine-grained
controls are needed of the input in a particular screen.

Implemented button masking in match screen. Only ROT button is
functional and the rest are ignored.